[Refactor] Book 어댑터 RestTemplate 이용하도록 수정
